+++ src/tensor_parallel/slicing_configs.py
@@ -21,8 +21,8 @@ from tensor_parallel.aux_actions import (
     split_num_heads,
 )
 from tensor_parallel.communications import CollectiveOperation
+from tensor_parallel.config import Config
 from tensor_parallel.per_device_tensors import PerDeviceTensors
-from tensor_parallel.slicer_wrapper import Config
 from tensor_parallel.state_actions import Scale, Split, SplitInChunks, StateAction
 from tensor_parallel.utils import nested_map
