+++ src/tensor_parallel/tensor_parallel.py
@@ -14,8 +14,9 @@ from torch.cuda.amp import autocast
 from torch.nn.parallel import parallel_apply
 
 from tensor_parallel.autoconfig import get_default_config
+from tensor_parallel.config import TENSOR_PARALLEL_USE_NATIVE, Config
 from tensor_parallel.cross_device_ops import broadcast_coalesced
-from tensor_parallel.slicer_wrapper import TENSOR_PARALLEL_USE_NATIVE, Config
+from tensor_parallel.shard import make_shard
 from tensor_parallel.utils import nested_flatten, nested_pack
 
 logger = logging.getLogger(__file__)
@@ -71,9 +72,7 @@ class TensorParallel(nn.Module):
         for rank, device in enumerate(self.devices):
             if delay_init:
                 device = torch.device("cpu")
-            self.module_shards.append(
-                tensor_parallel_config.make_shard(module, device, config_with_ops, rank=rank, world_size=world_size)
-            )
+            self.module_shards.append(make_shard(module, device, config_with_ops, rank=rank, world_size=world_size))
 
         # self-diagnostics: check if the model was sharded properly
